+++ config/routes.rb
@@ -15,11 +15,18 @@ RedmineApp::Application.routes.draw do
       
     # auto completes for hints for description
     # match 'auto_completes/attachment', :to => 'auto_completes#attachment_descriptions', :via => :get, :as => 'auto_complete_attachment_descriptions'
+    # match 'ttw_cats/:ttw_cat_id/ttw_cat/move/:id', to: 'ttw_cats#move_to', via: [:get, :post, :patch, :put]
 
     resources :ttw_cats, only: %i[load preview list_themes] do
       get 'load', on: :collection
       post 'list_themes', on: :collection
     end
+
+    # resources :ttw_cats, :except => :show  do
+    #   collection do
+    #     #match 'fields', :via => [:get, :post]
+    #   end
+    # end
 
 
     resources :ttw_issue_templates, :except => :show